refactor gs/blob file system
